+++ templates/java/CosmosDBTrigger/src/main/java/com/function/CosmosDBTriggerFunction.java
@@ -1,15 +1,10 @@
 package com.function;
 
 import com.microsoft.azure.functions.ExecutionContext;
-import com.microsoft.azure.functions.HttpMethod;
-import com.microsoft.azure.functions.HttpRequestMessage;
-import com.microsoft.azure.functions.HttpResponseMessage;
-import com.microsoft.azure.functions.HttpStatus;
-import com.microsoft.azure.functions.annotation.CosmosDBInput;
+import com.microsoft.azure.functions.annotation.CosmosDBTrigger;
 import com.microsoft.azure.functions.annotation.FunctionName;
-import com.microsoft.azure.functions.annotation.HttpTrigger;
 
-import java.util.Optional;
+import java.util.List;
 
 public class CosmosDBTriggerFunction {
 
@@ -23,9 +18,9 @@ public class CosmosDBTriggerFunction {
             connection = "CosmosDbConnection",
             createLeaseContainerIfNotExists = true
         )
-        Object inputItem,
+        List<Object> items,
         final ExecutionContext context
     ) {
-        context.getLogger().info("Items modified: " + inputItems.size());
+        context.getLogger().info("Items modified: " + items.size());
     }
 }